Moved all input setup into the "ConsumeController" to avoid the superclass problem. Load Level now needs a input to change to. Changed "ConsumeControl" to "IngestControl" (which are player attacks) to avoid confusion with "ConsumeController" (which handles player input).
